+++ src/components/film-details.js
@@ -1,47 +1,15 @@
 import {MONTH_NAMES} from '../data';
 import AbstractSmartComponent from "./abstract-smart-component";
 
-const IMG_HEIGHT = 55;
-const IMG_WIDTH = 55;
-
 const createGenresList = (genres) => {
   return genres.map((genre) => {
     return `<span class="film-details__genre">${genre}</span>`;
   }).join(` `);
 };
 
-const createComment = (commentsInf) => {
-  return commentsInf.map((comment) => {
-    const year = comment.date.getFullYear();
-    const month = comment.date.getUTCMonth() + 1;
-    const day = comment.date.getDate();
-    const hour = comment.date.getHours() + 1;
-    const minutes = comment.date.getMinutes() + 1;
-
-    return (`
-      <li class="film-details__comment">
-        <span class="film-details__comment-emoji">
-          <img src="./images/emoji/${comment.emotion}.png" width="55" height="55" alt="${comment.emotion}">
-        </span>
-        <div>
-          <p class="film-details__comment-text">${comment.text}</p>
-          <p class="film-details__comment-info">
-            <span class="film-details__comment-author">${comment.author}</span>
-            <span class="film-details__comment-day">${year}/${month}/${day} ${hour}:${minutes}</span>
-            <button class="film-details__comment-delete">Delete</button>
-          </p>
-        </div>
-      </li>
-    `);
-  }).join(`\n`);
-};
-
-const createFilmDetails = (film, options = {}) => {
+const createFilmDetails = (film) => {
   const {poster} = film;
-  const {emoji, emojiId} = options;
   const alt = poster.split(`.`)[0];
-  const comments = film.comments;
-  const commentsQuantity = film.comments.length;
   const writers = film.writers.join(`, `);
   const actors = film.actors.join(`, `);
   const description = film.description.join(` `);
@@ -50,7 +18,6 @@ const createFilmDetails = (film, options = {}) => {
   const fiilDay = film.date.getDate();
   const fiilDate = `${fiilDay} ${MONTH_NAMES[filmMonth]} ${filmYear}`;
 
-  const commentsMarkup = createComment(comments);
   const genresMarkup = createGenresList(film.genres);
 
   return (
@@ -130,45 +97,7 @@ const createFilmDetails = (film, options = {}) => {
         </div>
 
         <div class="form-details__bottom-container">
-          <section class="film-details__comments-wrap">
-            <h3 class="film-details__comments-title">Comments <span class="film-details__comments-count">${commentsQuantity}</span></h3>
 
-            <ul class="film-details__comments-list">
-              ${commentsMarkup}
-            </ul>
-
-            <div class="film-details__new-comment">
-              <div for="add-emoji" class="film-details__add-emoji-label">
-                ${emoji !== null ? emoji : ``}
-              </div>
-
-              <label class="film-details__comment-label">
-                <textarea class="film-details__comment-input" placeholder="Select reaction below and write comment here" name="comment"></textarea>
-              </label>
-
-              <div class="film-details__emoji-list">
-                <input ${emojiId === `emoji-smile` ? `checked` : ``} class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-smile" value="smile">
-                <label class="film-details__emoji-label" for="emoji-smile">
-                  <img src="./images/emoji/smile.png" width="30" height="30" alt="emoji">
-                </label>
-
-                <input ${emojiId === `emoji-sleeping` ? `checked` : ``} class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-sleeping" value="sleeping">
-                <label class="film-details__emoji-label" for="emoji-sleeping">
-                  <img src="./images/emoji/sleeping.png" width="30" height="30" alt="emoji">
-                </label>
-
-                <input ${emojiId === `emoji-puke` ? `checked` : ``} class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-puke" value="puke">
-                <label class="film-details__emoji-label" for="emoji-puke">
-                  <img src="./images/emoji/puke.png" width="30" height="30" alt="emoji">
-                </label>
-
-                <input ${emojiId === `emoji-angry` ? `checked` : ``} class="film-details__emoji-item visually-hidden" name="comment-emoji" type="radio" id="emoji-angry" value="angry">
-                <label class="film-details__emoji-label" for="emoji-angry">
-                  <img src="./images/emoji/angry.png" width="30" height="30" alt="emoji">
-                </label>
-              </div>
-            </div>
-          </section>
         </div>
       </form>
     </section>`
@@ -180,84 +109,37 @@ export default class FilmDetails extends AbstractSmartComponent {
     super();
     this._film = film;
 
-    this._closeDetailClick = null;
-    this._watchListClic = null;
-    this._historyClic = null;
-    this._favoriteClic = null;
-    this._emoji = null;
-    this._emojiId = null;
-    this._setEmojiClick();
-  }
-
-  _setEmojiClick() {
-    const element = this.getElement();
-
-    element.querySelector(`.film-details__emoji-list`)
-    .addEventListener(`click`, (evt) => {
-      evt.preventDefault();
-      const src = evt.target.attributes.src.value;
-      const alt = evt.target.alt;
-      const emojiId = evt.target.parentNode.attributes.for.value;
-      const img = `<img src=${src} alt=${alt} width=${IMG_WIDTH} height=${IMG_HEIGHT}>`;
-      this._emoji = img;
-      this._emojiId = emojiId;
-
-      this.rerender();
-    });
+    this.setCloseDetailClick(this.closeHandler);
   }
 
   rerender() {
     super.rerender();
   }
 
-  reset() {
-    this._emoji = null;
-    this._emojiId = null;
-
-    this.rerender();
-  }
-
-  recoveryListeners() {
-    this.setCloseDetailClick(this._closeDetailClick);
-    this.setWatchListClickHandler(this._watchListClic);
-    this.setHistoryClickHandler(this._historyClic);
-    this.setFavoriteClickHandler(this._favoriteClic);
-    this._setEmojiClick();
-  }
-
   getTemplate() {
-    return createFilmDetails(this._film, {
-      emoji: this._emoji,
-      emojiId: this._emojiId
-    });
+    return createFilmDetails(this._film);
   }
 
   setCloseDetailClick(handler) {
     const filmClose = this.getElement().querySelector(`.film-details__close-btn`);
     filmClose.addEventListener(`click`, handler);
 
-    this._closeDetailClick = handler;
+    this.closeHandler = handler;
   }
 
 
   setWatchListClickHandler(handler) {
     const watchlist = this.getElement().querySelector(`.film-details__control-label--watchlist`);
     watchlist.addEventListener(`click`, handler);
-
-    this._watchListClic = handler;
   }
 
   setHistoryClickHandler(handler) {
     const watched = this.getElement().querySelector(`.film-details__control-label--watched`);
     watched.addEventListener(`click`, handler);
-
-    this._historyClic = handler;
   }
 
   setFavoriteClickHandler(handler) {
     this.getElement().querySelector(`.film-details__control-label--favorite`)
       .addEventListener(`click`, handler);
-
-    this._favoriteClic = handler;
   }
 }
