+++ src/controllers/movie-controller.js
@@ -1,19 +1,19 @@
 import FilmCardComponent from "../components/film-card";
 import FilmDetailsComponent from "../components/film-details";
+import CommentsComponent from "../components/comments";
 import {render, replace, remove} from "../utils/render.js";
 
-export const Mode = {
+const Mode = {
   CLOSE: `close`,
   OPEN: `open`,
 };
-
-export const EmptyComment = {};
 
 export default class MovieController {
   constructor(container, onDataChange, onViewChange) {
     this._container = container;
     this._filmCardComponent = null;
     this._filmDetailsComponent = null;
+    this._commentsComponent = null;
     this._mode = Mode.CLOSE;
 
     this._onViewChange = onViewChange;
@@ -29,9 +29,11 @@ export default class MovieController {
   render(film) {
     const oldfilmCardComponent = this._filmCardComponent;
     const oldfilmDetailsComponent = this._filmDetailsComponent;
+    const oldcommentsComponent = this._commentsComponent;
 
     this._filmCardComponent = new FilmCardComponent(film);
     this._filmDetailsComponent = new FilmDetailsComponent(film);
+    this._commentsComponent = new CommentsComponent(film.comments);
 
     this._filmCardComponent.setPosterClickHandler((evt) => {
       evt.preventDefault();
@@ -87,9 +89,10 @@ export default class MovieController {
       }));
     });
 
-    if (oldfilmCardComponent && oldfilmDetailsComponent) {
+    if (oldfilmCardComponent && oldfilmDetailsComponent && oldcommentsComponent) {
       replace(this._filmCardComponent, oldfilmCardComponent);
       replace(this._filmDetailsComponent, oldfilmDetailsComponent);
+      replace(this._commentsComponent, oldcommentsComponent);
     } else {
       render(this._container, this._filmCardComponent);
     }
@@ -105,11 +108,13 @@ export default class MovieController {
     this._onViewChange();
     this._mode = Mode.OPEN;
     render(document.body, this._filmDetailsComponent);
+    const commentsContainer = this._filmDetailsComponent.getElement().querySelector(`.form-details__bottom-container`);
+    render(commentsContainer, this._commentsComponent);
   }
 
   _onCloseDetailClick() {
     this._mode = Mode.CLOSE;
-    this._filmDetailsComponent.reset();
+    this._commentsComponent.reset();
     remove(this._filmDetailsComponent);
   }
 }
